Separate out common form related utils
